refactor: combine ReleaseSourceRepo and multiReleaseSource
This is part of an "inline to refactor". I'd like to refactor the interfaces.
It also reduces confusion on what release source composite type to use.
